1-grid-world/requirements.txt
Keras>=2.11.0
numpy>=1.24.1
pandas>=1.5.3
matplotlib>=3.6.3
tensorflow>=2.11.0
Pillow>=9.4.0
gym>=0.26.2
h5py
scikit-image>=0.19.2